pyproject.toml
[project]
name = "wenah"
version = "0.1.0"
description = "Civil Rights Compliance Framework for AI/Software Products"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Project Wenah Team"}
]
keywords = ["civil-rights", "compliance", "ai", "machine-learning", "legal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # LLM & RAG
    "anthropic>=0.40.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-chroma>=0.2.0",

    # Vector Database & Embeddings
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",

    # Rule Engine & Data Processing
    "scikit-learn>=1.4.0",
    "PyYAML>=6.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # Configuration & Environment
    "python-dotenv>=1.0.0",

    # Async & HTTP
    "httpx>=0.27.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "black>=24.1.0",
    "pre-commit>=3.6.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[project.scripts]
wenah = "wenah.api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wenah"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
